+++ payments-core/src/main/java/com/stripe/android/model/StripeIntent.kt
@@ -263,6 +263,7 @@ sealed interface StripeIntent : StripeModel {
             override fun hashCode(): Int {
                 return 0
             }
+
             override fun equals(other: Any?): Boolean {
                 return this === other
             }