+++ payments-core/src/main/java/com/stripe/android/model/parsers/PaymentIntentJsonParser.kt
@@ -6,12 +6,15 @@ import com.stripe.android.core.model.StripeJsonUtils.optString
 import com.stripe.android.core.model.parsers.ModelJsonParser
 import com.stripe.android.core.model.parsers.ModelJsonParser.Companion.jsonArrayToList
 import com.stripe.android.model.Address
+import com.stripe.android.model.LuxeNextActionRepository
 import com.stripe.android.model.PaymentIntent
 import com.stripe.android.model.StripeIntent
 import org.json.JSONObject
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-class PaymentIntentJsonParser : ModelJsonParser<PaymentIntent> {
+class PaymentIntentJsonParser(
+    val luxeNextActionRepository: LuxeNextActionRepository = LuxeNextActionRepository.Instance
+) : ModelJsonParser<PaymentIntent> {
     override fun parse(json: JSONObject): PaymentIntent? {
         val objectType = optString(json, FIELD_OBJECT)
         if (OBJECT_TYPE != objectType) {
@@ -65,7 +68,19 @@ class PaymentIntentJsonParser : ModelJsonParser<PaymentIntent> {
             ShippingJsonParser().parse(it)
         }
         val nextActionData = json.optJSONObject(FIELD_NEXT_ACTION)?.let {
-            NextActionDataParser().parse(it)
+            when (
+                val luxeNextActionResult = luxeNextActionRepository.getAction(
+                    paymentMethod?.code,
+                    status,
+                    json
+                )
+            ) {
+                is LuxeNextActionRepository.Result.Action -> luxeNextActionResult.nextActionData
+                is LuxeNextActionRepository.Result.NoAction -> null
+                is LuxeNextActionRepository.Result.NotSupported -> {
+                    NextActionDataParser().parse(it)
+                }
+            }
         }
 
         val unactivatedPaymentMethods = jsonArrayToList(