+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * [ADDED][5340](https://github.com/stripe/stripe-android/pull/5340) Add a `resetCustomer` method to 
   `PaymentSheet`, that clears any persisted authentication state.
+* [FIXED][5388](https://github.com/stripe/stripe-android/pull/5388) Fixed issue with Appearance API
+  not working with `FlowController`
 
 ## 20.8.0 - 2022-08-01
 
@@ -30,6 +32,9 @@ and adds new `rememberLauncher` features for Payments
   Compose.
 * [DEPRECATED][5274](https://github.com/stripe/stripe-android/pull/5274)
   Deprecate `PaymentLauncher.createForCompose` in favor of  `PaymentLauncher.rememberLauncher`.
+
+### Identity
+* [ADDED][5370](https://github.com/stripe/stripe-android/pull/5370) Add factory method for Compose
 
 ## 20.7.0 - 2022-07-06